Tighten up repeated code, update README
Use some loops. This moves the initialization of the cybergear objects
into the setup() function which may or may not be appropriate. That's
necessary as c++ doesn't allow for control structures on the top
level.

Put steps in the README for getting this working on MacOS so I don't
have to think the next time I run into it.